--- _integration_tests/test/kpi/kpi_api_get_runtime_informations.js
@@ -1,116 +1,0 @@
-'use strict';
-
-const should = require('should');
-const TestFixtureProvider = require('../../dist/commonjs').TestFixtureProvider;
-
-describe('Management API -> Get ProcessInstance Runtime Informations - ', () => {
-
-  let testFixtureProvider;
-
-  let defaultIdentity;
-
-  const processModelId = 'kpi_api_test_data';
-
-  before(async () => {
-    testFixtureProvider = new TestFixtureProvider();
-    await testFixtureProvider.initializeAndStart();
-
-    defaultIdentity = testFixtureProvider.identities.defaultUser;
-  });
-
-  after(async () => {
-    await testFixtureProvider.tearDown();
-  });
-
-  it('should successfully get the runtime informations for a ProcessModel', async () => {
-    const runtimeInfos = await testFixtureProvider
-      .managementApiClient
-      .getRuntimeInformationForProcessModel(defaultIdentity, processModelId);
-
-    should(runtimeInfos).be.an.Array();
-    should(runtimeInfos).have.a.lengthOf(10, `Expected 10 runtime informations, but got ${runtimeInfos.length}.`);
-
-    const expectedFlowNodeIds = [
-      'StartEvent_1mox3jl',
-      'ExclusiveGateway_0fi1ct7',
-      'ScriptTask_1',
-      'ServiceTask_1',
-      'ExclusiveGateway_134ybqm',
-      'ParallelSplitGateway_1',
-      'ParallelJoinGateway_1',
-      'UserTask_1',
-      'UserTask_2',
-      'EndEvent_0eie6q6',
-    ];
-
-    for (const flowNodeId of expectedFlowNodeIds) {
-
-      const matchingRuntimeInfo = runtimeInfos.find((runtimeInfo) => {
-        return runtimeInfo.flowNodeId === flowNodeId;
-      });
-
-      should.exist(matchingRuntimeInfo);
-      should(matchingRuntimeInfo.processModelId).be.equal('kpi_api_test_data');
-      should(matchingRuntimeInfo).have.property('minRuntimeInMs');
-      should(matchingRuntimeInfo).have.property('maxRuntimeInMs');
-      should(matchingRuntimeInfo).have.property('arithmeticMeanRuntimeInMs');
-      should(matchingRuntimeInfo).have.property('firstQuartileRuntimeInMs');
-      should(matchingRuntimeInfo).have.property('medianRuntimeInMs');
-      should(matchingRuntimeInfo).have.property('thirdQuartileRuntimeInMs');
-    }
-  });
-
-  it('should successfully get the runtime information for a FlowNode with an odd number of executions', async () => {
-    const flowNodeToQuery = 'UserTask_1'; // in the metrics sample file, this task was run 5 times.
-    const runtimeInfo = await testFixtureProvider
-      .managementApiClient
-      .getRuntimeInformationForFlowNode(defaultIdentity, processModelId, flowNodeToQuery);
-
-    should(runtimeInfo).not.be.an.Array();
-
-    const expectedFlowNodeId = 'UserTask_1';
-    const expectedProcessModelId = 'kpi_api_test_data';
-    const expectedMinRuntimeInMs = 793;
-    const expectedMaxRuntimeInMs = 827;
-    const expectedArithmeticMeanRuntimeInMs = 812;
-    const expectedFirstQuartileRuntimeInMs = 793;
-    const expectedMedianRuntimeInMs = 814;
-    const expectedThirdQuartileRuntimeInMs = 827;
-
-    should(runtimeInfo.flowNodeId).be.equal(expectedFlowNodeId);
-    should(runtimeInfo.processModelId).be.equal(expectedProcessModelId);
-    should(runtimeInfo.minRuntimeInMs).be.equal(expectedMinRuntimeInMs);
-    should(runtimeInfo.maxRuntimeInMs).be.equal(expectedMaxRuntimeInMs);
-    should(runtimeInfo.arithmeticMeanRuntimeInMs).be.equal(expectedArithmeticMeanRuntimeInMs);
-    should(runtimeInfo.firstQuartileRuntimeInMs).be.equal(expectedFirstQuartileRuntimeInMs);
-    should(runtimeInfo.medianRuntimeInMs).be.equal(expectedMedianRuntimeInMs);
-    should(runtimeInfo.thirdQuartileRuntimeInMs).be.equal(expectedThirdQuartileRuntimeInMs);
-  });
-
-  it('should successfully get the runtime information for a FlowNode with an even number of executions', async () => {
-    const flowNodeToQuery = 'ScriptTask_1'; // in the metrics sample file, this task was run 20 times.
-    const runtimeInfo = await testFixtureProvider
-      .managementApiClient
-      .getRuntimeInformationForFlowNode(defaultIdentity, processModelId, flowNodeToQuery);
-
-    should(runtimeInfo).not.be.an.Array();
-
-    const expectedFlowNodeId = 'ScriptTask_1';
-    const expectedProcessModelId = 'kpi_api_test_data';
-    const expectedMinRuntimeInMs = 9;
-    const expectedMaxRuntimeInMs = 13;
-    const expectedArithmeticMeanRuntimeInMs = 10;
-    const expectedFirstQuartileRuntimeInMs = 9;
-    const expectedMedianRuntimeInMs = 10;
-    const expectedThirdQuartileRuntimeInMs = 12;
-
-    should(runtimeInfo.flowNodeId).be.equal(expectedFlowNodeId);
-    should(runtimeInfo.processModelId).be.equal(expectedProcessModelId);
-    should(runtimeInfo.minRuntimeInMs).be.equal(expectedMinRuntimeInMs);
-    should(runtimeInfo.maxRuntimeInMs).be.equal(expectedMaxRuntimeInMs);
-    should(runtimeInfo.arithmeticMeanRuntimeInMs).be.equal(expectedArithmeticMeanRuntimeInMs);
-    should(runtimeInfo.firstQuartileRuntimeInMs).be.equal(expectedFirstQuartileRuntimeInMs);
-    should(runtimeInfo.medianRuntimeInMs).be.equal(expectedMedianRuntimeInMs);
-    should(runtimeInfo.thirdQuartileRuntimeInMs).be.equal(expectedThirdQuartileRuntimeInMs);
-  });
-});
